+++ refacFabela/src/main/java/com/refacFabela/repository/TwSaldosRepository.java
@@ -1,5 +1,6 @@
 package com.refacFabela.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -13,7 +14,7 @@ import com.refacFabela.model.TwVentaProductoCancela;
 public interface TwSaldosRepository extends JpaRepository<TwVentaProductoCancela, Long> {
 
 	@Query(value="Select sum(n_precioPartida) from tw_venta_producto_cancela where n_idCaja=:id", nativeQuery = true) 
-	public Double totalCancela(Long id);
+	public BigDecimal totalCancela(Long id);
 	
 	@Query("Select e from TwVentaProductoCancela e where e.nIdProductos=:id order by e.dFecha desc ") 
 	public List<TwVentaProductoCancela> productosCancelados(Long id);
